Split project into 2 partls (app + library) was a bad idea.
- No watch on lib
- Can't move three examples into library, as resolt no benefit of split
