+++ angular.json
@@ -120,46 +120,6 @@
           }
         }
       }
-    },
-    "three-wrapper": {
-      "root": "projects/three-wrapper",
-      "sourceRoot": "projects/three-wrapper/src",
-      "projectType": "library",
-      "prefix": "lib",
-      "architect": {
-        "build": {
-          "builder": "@angular-devkit/build-ng-packagr:build",
-          "options": {
-            "tsConfig": "projects/three-wrapper/tsconfig.lib.json",
-            "project": "projects/three-wrapper/ng-package.json"
-          },
-          "configurations": {
-            "production": {
-              "project": "projects/three-wrapper/ng-package.prod.json"
-            }
-          }
-        },
-        "test": {
-          "builder": "@angular-devkit/build-angular:karma",
-          "options": {
-            "main": "projects/three-wrapper/src/test.ts",
-            "tsConfig": "projects/three-wrapper/tsconfig.spec.json",
-            "karmaConfig": "projects/three-wrapper/karma.conf.js"
-          }
-        },
-        "lint": {
-          "builder": "@angular-devkit/build-angular:tslint",
-          "options": {
-            "tsConfig": [
-              "projects/three-wrapper/tsconfig.lib.json",
-              "projects/three-wrapper/tsconfig.spec.json"
-            ],
-            "exclude": [
-              "**/node_modules/**"
-            ]
-          }
-        }
-      }
     }
   },
   "defaultProject": "angular-template-for-dummies"
