--- projects/three-wrapper/src/public_api.ts
@@ -1,19 +1,0 @@
-/*
- * Public API Surface of three-wrapper
- */
-
-export * from './lib/three-wrapper.module';
-
-export * from './lib/cameras/abstract-camera';
-export * from './lib/cameras/perspective-camera.directive';
-
-export * from './lib/renderer/renderer.component';
-
-export * from './lib/objects/abstract-object-3d';
-export * from './lib/objects/axes-helper.directive';
-export * from './lib/objects/scene.directive';
-export * from './lib/objects/model-loader.directive';
-export * from './lib/objects/grid-helper.directive';
-export * from './lib/objects/object-loader.directive';
-export * from './lib/objects/point-light.directive';
-
