+++ src/app/three-examples/objects/abstract-object-3d.ts
