+++ src/app/three-examples/objects/model-loader.directive.ts
