+++ src/app/three-examples/objects/object-loader.directive.ts
