+++ src/app/three-examples/objects/point-light.directive.ts
