+++ src/app/three-examples/objects/scene.directive.ts
