+++ src/app/three-examples/pipes/degrees-to-radians.pipe.ts
