+++ src/app/three-examples/renderer/renderer.component.ts
