+++ tsconfig.json
@@ -15,11 +15,6 @@
     "lib": [
       "es2017",
       "dom"
-    ],
-    "paths": {
-      "three-wrapper": [
-        "dist/three-wrapper"
-      ]
-    }
+    ]
   }
 }
